Image improvements
The outline for the images was weird and had "gaps" in a few places. Now, the outline should look better. On the other hand, every image sent was a JPEG and the quality for the text wasn't the best, now images are in PNG format and that should improve the quality enough to improve readability. Lastly, the size of the font was increased to help read the text a bit more :godmode: